requirements.txt
# Core
google-genai>=1.0.0          # New unified Gemini SDK
pydantic>=2.7.0
python-dotenv>=1.0.1

# UI
streamlit>=1.36.0
plotly>=5.22.0
streamlit-timeline>=0.0.5

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Parsing & forensics
python-dateutil>=2.9.0
pyyaml>=6.0.1
stix2>=3.0.1

# Eval
scikit-learn>=1.5.0
tabulate>=0.9.0

# Dev
pytest>=8.2.0
pytest-cov>=5.0.0
ruff>=0.5.0
